[RENAME] bookAPI -> book (책관련 코드 관리)
