+++ src/main/java/com/mkl/book/book/controller/ReviewController.java
@@ -1,4 +1,4 @@
-package com.mkl.book.bookAPI.controller;
+package com.mkl.book.book.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
