+++ src/main/java/com/mkl/book/book/repository/BookRepository.java
@@ -1,6 +1,6 @@
-package com.mkl.book.bookAPI.repository;
+package com.mkl.book.book.repository;
 
-import com.mkl.book.bookAPI.domain.Book;
+import com.mkl.book.book.domain.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
