+++ src/main/java/com/mkl/book/book/service/BookDBService.java
@@ -1,7 +1,7 @@
-package com.mkl.book.bookAPI.service;
+package com.mkl.book.book.service;
 
-import com.mkl.book.bookAPI.domain.Book;
-import com.mkl.book.bookAPI.repository.BookRepository;
+import com.mkl.book.book.domain.Book;
+import com.mkl.book.book.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
